Fix for #49
* Added check to make sure exp.ExpressionList is not empty to sql_dialect
