+++ exp/exp_map.go
@@ -1,5 +1,12 @@
 package exp
 
+import (
+	"sort"
+	"strings"
+
+	"github.com/doug-martin/goqu/v7/internal/errors"
+)
+
 type (
 	// A map of expressions to be ANDed together where the keys are string that will be used as Identifiers and values
 	// will be used in a boolean operation.
@@ -27,6 +34,10 @@ func (e Ex) Clone() Expression {
 	return ret
 }
 
+func (e Ex) IsEmpty() bool {
+	return len(e) == 0
+}
+
 func (e Ex) ToExpressions() (ExpressionList, error) {
 	return mapToExpressionList(e, AndType)
 }
@@ -43,6 +54,102 @@ func (eo ExOr) Clone() Expression {
 	return ret
 }
 
+func (eo ExOr) IsEmpty() bool {
+	return len(eo) == 0
+}
+
 func (eo ExOr) ToExpressions() (ExpressionList, error) {
 	return mapToExpressionList(eo, OrType)
+}
+
+func getExMapKeys(ex map[string]interface{}) []string {
+	var keys []string
+	for key := range ex {
+		keys = append(keys, key)
+	}
+	sort.Strings(keys)
+	return keys
+}
+
+func mapToExpressionList(ex map[string]interface{}, eType ExpressionListType) (ExpressionList, error) {
+	keys := getExMapKeys(ex)
+	ret := make([]Expression, len(keys))
+	for i, key := range keys {
+		lhs := ParseIdentifier(key)
+		rhs := ex[key]
+		var exp Expression
+		if op, ok := rhs.(Op); ok {
+			ors, err := createOredExpressionFromMap(lhs, op)
+			if err != nil {
+				return nil, err
+			}
+			exp = NewExpressionList(OrType, ors...)
+		} else {
+			exp = lhs.Eq(rhs)
+		}
+		ret[i] = exp
+	}
+	if eType == OrType {
+		return NewExpressionList(OrType, ret...), nil
+	}
+	return NewExpressionList(AndType, ret...), nil
+}
+
+func createOredExpressionFromMap(lhs IdentifierExpression, op Op) ([]Expression, error) {
+	opKeys := getExMapKeys(op)
+	ors := make([]Expression, len(opKeys))
+	for j, opKey := range opKeys {
+		if exp, err := createExpressionFromOp(lhs, opKey, op); err != nil {
+			return nil, err
+		} else if exp != nil {
+			ors[j] = exp
+		}
+	}
+	return ors, nil
+}
+
+func createExpressionFromOp(lhs IdentifierExpression, opKey string, op Op) (exp Expression, err error) {
+	switch strings.ToLower(opKey) {
+	case "eq":
+		exp = lhs.Eq(op[opKey])
+	case "neq":
+		exp = lhs.Neq(op[opKey])
+	case "is":
+		exp = lhs.Is(op[opKey])
+	case "isnot":
+		exp = lhs.IsNot(op[opKey])
+	case "gt":
+		exp = lhs.Gt(op[opKey])
+	case "gte":
+		exp = lhs.Gte(op[opKey])
+	case "lt":
+		exp = lhs.Lt(op[opKey])
+	case "lte":
+		exp = lhs.Lte(op[opKey])
+	case "in":
+		exp = lhs.In(op[opKey])
+	case "notin":
+		exp = lhs.NotIn(op[opKey])
+	case "like":
+		exp = lhs.Like(op[opKey])
+	case "notlike":
+		exp = lhs.NotLike(op[opKey])
+	case "ilike":
+		exp = lhs.ILike(op[opKey])
+	case "notilike":
+		exp = lhs.NotILike(op[opKey])
+	case "between":
+		rangeVal, ok := op[opKey].(RangeVal)
+		if ok {
+			exp = lhs.Between(rangeVal)
+		}
+	case "notbetween":
+		rangeVal, ok := op[opKey].(RangeVal)
+		if ok {
+			exp = lhs.NotBetween(rangeVal)
+		}
+	default:
+		err = errors.New("unsupported expression type %s", op)
+	}
+	return exp, err
 }
